update: refactor minify-selectors to postcss 8
